requirements.txt
pillow
opencv-python
requests
tqdm
numpy
label-studio
timm
torch
torchvision
ultralytics
